pyproject.toml
[project]
name = "redaction-tool"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.3",
    "pillow>=12.2.0",
    "pymupdf>=1.27.2.2",
]
